Move candidate list filtering form and pagination to separate component
